requirements.txt
Django==1.9.6
Jinja2==2.8
MarkupSafe==0.23
Paste==2.0.3
PasteDeploy==1.5.2
PasteScript==2.0.2
Pillow==3.3.0
Twisted==16.2.0
XlsxWriter==0.9.3
amqp==1.4.9
anyjson==0.3.3
argparse==1.2.1
arrow==0.8.0
asgi-redis==0.13.1
asgiref==0.13.3
autobahn==0.14.1
billiard==3.3.0.23
celery==3.1.23
certifi==2016.8.2
channels==0.14.2
coreapi==1.32.0
daphne==0.12.1
django-braces==1.9.0
django-cleanup==0.4.2
django-crispy-forms==1.6.0
django-datatables-view==1.13.0
django-debug-toolbar==1.4
django-extra-views==0.8.0
django-guardian==1.4.4
django-mail-templated==2.6.2
django-rest-swagger==2.0.3
django-smart-selects==1.2.2
django-sslserver==0.19
django-tables2==1.2.5
django-telegram-bot==0.6.0
djangorestframework==3.3.3
et-xmlfile==1.0.1
future==0.15.2
image==1.5.3
isoweek==1.3.0
itypes==1.1.0
jdcal==1.2
jinja2-time==0.1.0
kik==1.0.7
kombu==3.0.35
messengerbot==0.1.4
msgpack-python==0.4.7
openapi-codec==1.0.0
openpyxl==2.3.5
permabots==2.2.3
protobuf==3.0.0
pycrypto==2.6.1
python-dateutil==2.5.3
python-telegram-bot==4.2.0
pytz==2016.4
redis==2.10.5
requests==2.10.0
shortuuid==0.4.3
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.19
txaio==2.5.1
uritemplate==0.6
urllib3==1.16
wsgiref==0.1.2
zope.interface==4.2.0
